Fix ADSB ICAO address type

+++ cFS/apps/ardupilot/fsw/src/apProcessMessages.c
@@ -773,7 +773,7 @@ void ARDUPILOT_ProcessPacket() {
 
             double heading = fmod(2 * M_PI + atan2(traffic->ve, traffic->vn), 2 * M_PI) * 180 / M_PI;
             double speed = sqrt(traffic->vn * traffic->vn + traffic->ve * traffic->ve);
-            mavlink_msg_adsb_vehicle_pack(sysid_ic, compid_ic, &msg, (uint8_t)traffic->index,
+            mavlink_msg_adsb_vehicle_pack(sysid_ic, compid_ic, &msg, traffic->index,
                                           (int32_t)(traffic->latitude * 1E7),
                                           (int32_t)(traffic->longitude * 1E7),
                                           ADSB_ALTITUDE_TYPE_GEOMETRIC,
